agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "python-dotenv>=1.2.2",
    "railtracks>=1.3.6",
    "railtracks-cli>=1.2.3",
    "uvicorn[standard]>=0.42.0",
    "websockets>=16.0",
]
